requirements.txt
chat-downloader==0.2.8
flask==2.3.2
gunicorn==21.2.0
numpy==1.24.2
pandas==2.0.3
plotly==5.15.0
pywebview==5.1
